DRVLIB_SW-853: DP export ctp_tc api and return 'inst' for bridge device

+++ drivers/net/ethernet/lantiq/datapath/datapath_instance.c
@@ -277,7 +277,7 @@ int dp_inst_add_dev(struct net_device *dev, char *subif_name, int inst,
 					 sizeof(*dev->netdev_ops),
 					 (void **)&dp_dev->old_dev_ops,
 					 &dp_dev->new_dev_ops,
-					 (unsigned long)&dp_ndo_setup_tc);
+					 &dp_ndo_setup_tc);
 			if (err)
 				return DP_FAILURE;
 		}
